back/requirements.txt
asgiref==3.7.2
certifi==2024.12.14
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.4.1
cryptography==44.0.0
distlib==0.3.9
Django==3.2.25
django-allauth==0.63.6
django-cors-headers==4.1.0
djangorestframework==3.15.1
dj-rest-auth==4.0.1
filelock==3.12.2
idna==3.10
importlib-metadata==6.7.0
oauthlib==3.2.2
Pillow==10.2.0
pipenv==2023.10.3
platformdirs==4.0.0
pycparser==2.21
PyJWT==2.8.0
pytz==2024.2
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
virtualenv==20.26.6
zipp==3.15.0
psycopg2-binary==2.9.9
daphne
djangorestframework-simplejwt
